+++ core/src/main/java/net/momirealms/craftengine/core/font/IllegalCharacterProcessResult.java
@@ -1,0 +1,12 @@
+package net.momirealms.craftengine.core.font;
+
+public record IllegalCharacterProcessResult(boolean has, String text) {
+
+    public static IllegalCharacterProcessResult has(String text) {
+        return new IllegalCharacterProcessResult(true, text);
+    }
+
+    public static IllegalCharacterProcessResult not() {
+        return new IllegalCharacterProcessResult(false, "");
+    }
+}
